+++ GameEngineVS/GameEngine/GameEngine/Quaternion.cpp
@@ -83,11 +83,11 @@ void Quaternion::normalize()
 
 void Quaternion::convertToUnitFormQuaternion()
 {
-	double angle = qw * M_PI / 180;
+	double angle = qw * M_PI / 180.f;
 	Vector3D v = Vector3D(qx,qy,qz);
 	v.Normalize();
-	qw *= cosf(angle * 0.5f);
-	v.scale(sinf(angle * 0.5f), sinf(angle * 0.5f), sinf(angle * 0.5f));
+	qw *= (float) cosf(angle * 0.5f);
+	v.scale((float) sinf(angle * 0.5f), (float) sinf(angle * 0.5f), sinf(angle * 0.5f));
 	qx = v.x;
 	qy = v.y; 
 	qz = v.z;
@@ -115,7 +115,7 @@ Quaternion Quaternion::conjugate()
 }
 
 Vector3D Quaternion::quaternionToVector(float angle) {
-	return Vector3D(qx*sin((M_PI*angle/180.f)/2.f), qy*sin((M_PI*angle / 180.f) / 2.f),qz*sin((M_PI*angle/180.f)/2.f),cos((M_PI*angle/180.f)/2.f));
+	return Vector3D((float) qx*sin((M_PI*angle/180.f)/2.f), (float) qy*sin((M_PI*angle / 180.f) / 2.f), (float) qz*sin((M_PI*angle/180.f)/2.f), (int) cos((M_PI*angle/180.f)/2.f));
 }
 
 Matrix4 Quaternion::quaternionToMatrix4()
